+++ roto-server/pages/master/baterai.vue
@@ -172,17 +172,34 @@ export default {
     },
     methods:{
         async deleteData(id){
-            const lokasi = this.$auth.user.lokasi
-            const resp = await axios.delete(`/server/master/deletebaterai/${id}/${lokasi}`)
-            if(resp){
-                this.baterais.splice(this.baterais.indexOf(id), 1);
-                this.$router.push('/master/baterai')
-                swal('data dihapus',{icon:'success'})
-            }
-            if(resp.data.errmsg){
-                this.$router.push('/master/ups')
-                swal("Error", resp.data.errmsg,{icon:'error'})
-            }
+            swal({
+                title: 'anda yakin?',
+                text: 'sekali dihapus, data tidak akan bisa kembali',
+                icon: 'warning',
+                buttons: true,
+                dangerMode: true
+            }).then(suc=>{
+                if(suc){
+                    const lokasi = this.$auth.user.lokasi
+                    axios.delete(`/server/master/deletebaterai/${id}/${lokasi}`)
+                    .then(resp=>{
+                        if(resp){
+                            this.baterais.splice(this.baterais.indexOf(id), 1);
+                            this.$router.push('/master/baterai')
+                            swal('data dihapus',{icon:'success'})
+                        }
+                    }).catch(err=>{
+                        if(err.data.errmsg){
+                            this.$router.push('/master/ups')
+                            swal("Error", err.data.errmsg,{icon:'error'})
+                        }
+                    })
+                } else{
+                    swal('Error','gagal menghapus',{icon:'error'})
+                }
+            }).catch(err=>{
+                swal('Error','ada yang salah',{icon:'error'})
+            })
         },
         
         async caribarangbaterai(){
