+++ roto-server/pages/master/keyboard.vue
@@ -175,19 +175,36 @@ export default {
     },
     methods:{
         async deleteData(id){
-            let indexOfArrayItem = this.keyboards.findIndex(i => i.id_keyboard === id)
+            swal({
+                title: 'anda yakin?',
+                text: 'sekali dihapus, data tidak akan bisa kembali',
+                icon: 'warning',
+                buttons: true,
+                dangerMode: true
+            }).then(suc=>{
+                if(suc){
+                    let indexOfArrayItem = this.keyboards.findIndex(i => i.id_keyboard === id)
 
-            const lokasi = this.$auth.user.lokasi
-            const resp = await axios.delete(`/server/master/deletekeyboard/${id}/${lokasi}`)
-            if(resp){
-                this.keyboards.splice(indexOfArrayItem, 1);
-                this.$router.push('/master/keyboard')
-                swal('data dihapus',{icon:'success'})
-            }
-            if(resp.data.errmsg){
-                this.$router.push('/master/keyboard')
-                swal("Error", resp.data.errmsg,{icon:'error'})
-            }
+                    const lokasi = this.$auth.user.lokasi
+                    axios.delete(`/server/master/deletekeyboard/${id}/${lokasi}`)
+                    .then(resp=>{
+                        if(resp){
+                            this.keyboards.splice(indexOfArrayItem, 1);
+                            this.$router.push('/master/keyboard')
+                            swal('data dihapus',{icon:'success'})
+                        }
+                    }).catch(err=>{
+                        if(err.data.errmsg){
+                            this.$router.push('/master/keyboard')
+                            swal("Error", err.data.errmsg,{icon:'error'})
+                        }
+                    })
+                }else{
+                    swal('Error','gagal menghapus',{icon:'error'})
+                }
+            }).catch(err=>{
+                swal('Error','ada yang salah',{icon:'error'})
+            })
         },
          async caribarangkeyboard(){
              this.carikeyboard = []
@@ -200,7 +217,8 @@ export default {
     async mounted(){
         try{
             const lokasi = this.$auth.user.lokasi
-            const resp = await axios.get(`http://localhost:3000/server/masterkeyboard?lokasi=${lokasi}`)
+            const idlogin = this.$auth.user.id
+            const resp = await axios.get(`http://localhost:3000/server/masterkeyboard/${lokasi}/${idlogin}`)
             resp.data.forEach(keyboard => {
                 this.keyboards.push(keyboard)
             })
