+++ roto-server/pages/master/mouse.vue
@@ -174,19 +174,36 @@ export default {
     },
     methods:{
         async deleteData(id){
-            let indexOfArrayItem = this.mouses.findIndex(i => i.id_mouse === id)
+            swal({
+                title: 'anda yakin?',
+                text: 'sekali dihapus, data tidak akan bisa kembali',
+                icon: 'warning',
+                buttons: true,
+                dangerMode: true
+            }).then(suc=>{
+                if(suc){
+                    let indexOfArrayItem = this.mouses.findIndex(i => i.id_mouse === id)
 
-            const lokasi = this.$auth.user.lokasi
-            const resp = await axios.delete(`/server/master/deletemouse/${id}/${lokasi}`)
-            if(resp){
-                this.mouses.splice(indexOfArrayItem, 1);
-                this.$router.push('/master/mouse')
-                swal('data dihapus',{icon:'success'})
-            }
-            if(resp.data.errmsg){
-                this.$router.push('/master/mouse')
-                swal("Error", resp.data.errmsg,{icon:'error'})
-            }
+                    const lokasi = this.$auth.user.lokasi
+                    axios.delete(`/server/master/deletemouse/${id}/${lokasi}`)
+                    .then(resp=>{
+                        if(resp){
+                            this.mouses.splice(indexOfArrayItem, 1);
+                            this.$router.push('/master/mouse')
+                            swal('data dihapus',{icon:'success'})
+                        }
+                    }).catch(err=>{
+                        if(err.data.errmsg){
+                            this.$router.push('/master/mouse')
+                            swal("Error", err.data.errmsg,{icon:'error'})
+                        }
+                    })
+                }else{
+                    swal('Error','gagal menghapus',{icon:'error'})
+                }
+            }).catch(err=>{
+                swal('Error','ada yang salah',{icon:'error'})
+            })
 
         },
         async caribarangmouse(){
@@ -200,7 +217,8 @@ export default {
     async mounted(){
         try{
             const lokasi = this.$auth.user.lokasi
-            const resp = await axios.get(`http://localhost:3000/server/mastermouse?lokasi=${lokasi}`)
+            const idlogin = this.$auth.user.id
+            const resp = await axios.get(`http://localhost:3000/server/mastermouse/${lokasi}/${idlogin}`)
             resp.data.forEach(mouse => {
             this.mouses.push(mouse)
         })
